+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-* `20/03/27` [opt] publish.gradle
+* `20/03/31` [add] Publish v1.27.0.
+* `20/03/30` [add] BatteryUtils.
+* `20/03/27` [add] publish.gradle.
+* `20/03/20` [add] UtilsBridge to clean the utils.
 * `20/01/17` [upd] Leak Canary to v2.1.
 * `19/11/30` [add] Publish bus plugin v2.4. Publish api plugin v1.2.
 * `19/11/28` [add] Publish v1.26.0.